+++ CHANGELOG.md
@@ -3,6 +3,17 @@ All notable changes to this project will be documented in this file.
  
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
+## [0.1.0] - 2023-10-06
+ 
+### Added
+- Made MemoizR async.
+- Added Structured Concurrency primitives
+- 
+   
+### Changed
+- 
+ 
+### Fixed
 
 ## [0.0.4] - 2023-09-18